Rewrite: Automate the entire process
Instead of requiring you to create a tileset and then creating a map from it that you can then turn into a tileset, the script now lets you pick the image, and takes care of creating the intermediate map (you can opt to save it as an image or a TMX) and creating the final tileset for you. This should streamline using RPG Maker tilesets in Tiled significantly, as it essentially lets you use them through a dialog very similar to the standard New Tileset dialog.

The script still doesn't automatically generate Terrains, I would like to add that eventually. It's tedious so I haven't gotten around to it :]

Because this version automates the creation of the intermediate tileset, if you want to use it in combination with Import Metatile Terrains, you'll need to change the order in which you do things: use this script first, save the intermediate as a TileMap, and then open that tilemap and edit its embedded Tileset.